fix vlad's comments

+++ core/src/main/java/io/questdb/griffin/engine/functions/math/LogIntFunctionFactory.java
@@ -65,7 +65,7 @@ public class LogIntFunctionFactory implements FunctionFactory {
         @Override
         public int getInt(Record rec) {
             int value = arg.getInt(rec);
-            return value != Numbers.INT_NaN ? (int) Math.log(value) : Double.NaN;
+            return value != Numbers.INT_NaN ? (int) Math.log(value) : (int) Double.NaN;
         }
     }
 }
