+++ core/src/main/java/io/questdb/griffin/engine/functions/math/SqrtIntFunctionFactory.java
@@ -65,7 +65,7 @@ public class SqrtIntFunctionFactory implements FunctionFactory {
         @Override
         public int getInt(Record rec) {
             int value = arg.getInt(rec);
-            return value != Numbers.INT_NaN ? (int) Math.sqrt(value) : Double.NaN;
+            return value != Numbers.INT_NaN ? (int) Math.sqrt(value) : (int) Double.NaN;
         }
     }
 }
