+++ core/src/main/java/io/questdb/griffin/engine/functions/math/SqrtLongFunctionFactory.java
@@ -51,7 +51,7 @@ public class SqrtLongFunctionFactory implements FunctionFactory {
         @Override
         public long getLong(Record rec) {
             long value = arg.getLong(rec);
-            return value != Numbers.LONG_NaN ? (long) Math.sqrt(value) : Double.NaN;
+            return value != Numbers.LONG_NaN ? (long) Math.sqrt(value) : (int) Double.NaN;
         }
     }
 }
